+++ hnnt-chat-web/src/components/Modal/Modal.js
@@ -18,6 +18,9 @@ import { useSelector } from 'react-redux';
 
 import PopupViewImage from '../Popup/PopupViewImage';
 import ModalAccount from './ModalAccount';
+import ModalAvatar from './ModalAvatar';
+import ModalEdit from './ModalEdit';
+import ModalEditAvatar from './ModalEditAvatar';
 
 function Modal({ isOpen, onClose }) {
     const userActive = useSelector((state) => state.auth.userActive);
@@ -119,287 +122,48 @@ function Modal({ isOpen, onClose }) {
                 </div>
 
                 {isType === 'avatar' ? (
-                    <motion.div
-                        initial={{ x: '100%', opacity: 0 }} // Bắt đầu ngoài màn hình bên phải
-                        animate={{ x: '0%', opacity: 1 }} // Trượt vào giữa màn hình
-                        exit={{ x: '100%', opacity: 0 }} // Khi đóng, trượt ngược ra phải
-                        transition={{ duration: 0.4, ease: 'easeInOut' }}
-                    >
-                        <div className="border-t pt-2 flex justify-center">
-                            <input
-                                type="file"
-                                accept="image/*"
-                                ref={fileInputRef}
-                                className="hidden"
-                                onChange={(e) => {
-                                    const file = e.target.files[0];
-                                    if (file) {
-                                        const imageUrl = URL.createObjectURL(file); // Tạo URL tạm thời
-                                        setImage(imageUrl);
-                                        setIsType('editAvatar');
-                                    }
-                                }}
-                            />
-                            <button
-                                onClick={handleTakePhoto}
-                                className="w-11/12 text-blue-800 bg-blue-200 hover:bg-blue-300 rounded-md flex items-center justify-center gap-2 p-2 cursor-pointer"
-                            >
-                                <MdOutlineInsertPhoto className="cursor-pointer" />
-                                <p className="font-semibold">Tải lên từ máy tính</p>
-                            </button>
-                        </div>
-
-                        <p className="font-semibold ml-4 mt-6">Ảnh đại diện của tôi</p>
-
-                        <div className="grid grid-cols-4 gap-4 w-full px-6 mt-3">
-                            <img
-                                src={Avatar}
-                                alt="avatar"
-                                className="w-16 h-16 object-cover rounded-full border-1 border-black cursor-pointer"
-                            />
-                            <img
-                                src={Avatar}
-                                alt="avatar"
-                                className="w-16 h-16 object-cover rounded-full border-1 border-black cursor-pointer"
-                            />
-                            <img
-                                src={Avatar}
-                                alt="avatar"
-                                className="w-16 h-16 object-cover rounded-full border-1 border-black cursor-pointer"
-                            />
-                            <img
-                                src={Avatar}
-                                alt="avatar"
-                                className="w-16 h-16 object-cover rounded-full border-1 border-black cursor-pointer"
-                            />
-                            <img
-                                src={Avatar}
-                                alt="avatar"
-                                className="w-16 h-16 object-cover rounded-full border-1 border-black cursor-pointer"
-                            />
-                        </div>
-                    </motion.div>
+                    <ModalAvatar setImage={setImage} setIsType={setIsType} />
                 ) : isType === 'edit' ? (
-                    <motion.div
-                        initial={{ x: '100%', opacity: 0 }} // Bắt đầu ngoài màn hình bên phải
-                        animate={{ x: '0%', opacity: 1 }} // Trượt vào giữa màn hình
-                        exit={{ x: '100%', opacity: 0 }} // Khi đóng, trượt ngược ra phải
-                        transition={{ duration: 0.4, ease: 'easeInOut' }}
-                        className="flex flex-col h-full"
-                    >
-                        <div className="border-t pt-2 mx-4">
-                            <p className="text-sm">Tên hiển thị</p>
-                            <input
-                                type="text"
-                                id="default-search"
-                                class="w-full border-2 border-gray-200 p-2 rounded-md mt-1 focus:ring-blue-500 focus:border-blue-500"
-                                required
-                                value={isName}
-                                onChange={(e) => setIsName(e.target.value)}
-                            />
-
-                            <p className="text-base font-semibold mt-6">Thông tin cá nhân</p>
-                            <div className="flex space-x-4 mt-4 ">
-                                {/* Giới tính */}
-                                <FormControl>
-                                    <RadioGroup
-                                        row
-                                        aria-labelledby="demo-row-radio-buttons-group-label"
-                                        name="row-radio-buttons-group"
-                                        value={gender}
-                                        onChange={(e) => setGender(e.target.value)}
-                                    >
-                                        <FormControlLabel
-                                            className="text-gray-700"
-                                            value="Nam"
-                                            control={<Radio />}
-                                            label="Nam"
-                                        />
-                                        <FormControlLabel
-                                            className="text-gray-700"
-                                            value="Nữ"
-                                            control={<Radio />}
-                                            label="Nữ"
-                                        />
-                                    </RadioGroup>
-                                </FormControl>
-                            </div>
-
-                            {/* ngày sinh */}
-                            <p className="text-sm mt-4 mb-2">Ngày sinh</p>
-                            <div className="flex gap-2 w-full h-10">
-                                {/* Ngày */}
-                                <Select
-                                    value={day}
-                                    onChange={(e) => setDay(e.target.value)}
-                                    style={{ minWidth: '30%' }}
-                                    MenuProps={{ PaperProps: { sx: { maxHeight: 230 } } }}
-                                    className="p-2 border rounded-md focus:ring-2 focus:ring-blue-500"
-                                >
-                                    {Array.from({ length: maxDays }, (_, i) => (
-                                        <MenuItem key={i + 1} value={i + 1}>
-                                            {i + 1}
-                                        </MenuItem>
-                                    ))}
-                                </Select>
-
-                                {/* Tháng */}
-                                <Select
-                                    value={month}
-                                    onChange={(e) => setMonth(e.target.value)}
-                                    style={{ minWidth: '30%' }}
-                                    MenuProps={{ PaperProps: { sx: { maxHeight: 230 } } }}
-                                    className="p-2 border rounded-md focus:ring-2 focus:ring-blue-500"
-                                >
-                                    {Array.from({ length: 12 }, (_, i) => (
-                                        <MenuItem key={i + 1} value={i + 1}>
-                                            {i + 1}
-                                        </MenuItem>
-                                    ))}
-                                </Select>
-
-                                {/* Năm */}
-                                <Select
-                                    value={year}
-                                    onChange={(e) => setYear(e.target.value)}
-                                    style={{ minWidth: '35%' }}
-                                    MenuProps={{ PaperProps: { sx: { maxHeight: 230 } } }}
-                                    className="p-2 border rounded-md focus:ring-2 focus:ring-blue-500"
-                                >
-                                    {Array.from({ length: 100 }, (_, i) => {
-                                        const y = currentYear - i;
-                                        return (
-                                            <MenuItem key={y} value={y}>
-                                                {y}
-                                            </MenuItem>
-                                        );
-                                    })}
-                                </Select>
-                            </div>
-                        </div>
-
-                        {/* button cập nhật */}
-                        <div className="border-t pt-2 flex justify-end mt-auto">
-                            <button
-                                onClick={() => setIsType('profile')}
-                                className="bg-gray-200 hover:bg-gray-300 rounded-md flex items-center justify-center gap-2 p-2 cursor-pointer mr-2"
-                            >
-                                <p className="font-semibold">Hủy</p>
-                            </button>
-
-                            <button
-                                onClick={onClose}
-                                className="bg-blue-600 hover:bg-blue-800 rounded-md flex items-center justify-center gap-2 p-2 cursor-pointer mr-3"
-                            >
-                                <p className="font-semibold text-white">Cập nhật</p>
-                            </button>
-                        </div>
-                    </motion.div>
+                    <ModalEdit setIsType={setIsType} onClose={onClose} />
                 ) : isType === 'editAvatar' ? (
                     image && (
-                        <>
-                            <div className="flex-1 overflow-auto pt-6 border-t border-gray-200 relative">
-                                <div className="w-full flex justify-center relative">
-                                    {/* Ảnh hiển thị */}
-                                    <img src={image} alt="Preview" className="mt-2 max-w-full h-80 object-cover " />
+                        <ModalEditAvatar image={image} setIsType={setIsType} onClose={onClose} />
+                        // <>
+                        //     <div className="flex-1 overflow-auto pt-6 border-t border-gray-200 relative">
+                        //         <div className="w-full flex justify-center relative">
+                        //             {/* Ảnh hiển thị */}
+                        //             <img src={image} alt="Preview" className="mt-2 max-w-full h-80 object-cover " />
 
-                                    {/* Lớp mờ với vùng tròn trong suốt */}
-                                    <div
-                                        className="absolute inset-0 bg-black/50"
-                                        style={{
-                                            WebkitMaskImage: `radial-gradient(circle 160px at center, transparent 160px, black 161px)`,
-                                            maskImage: `radial-gradient(circle 160px at center, transparent 160px, black 161px)`,
-                                        }}
-                                    />
-                                </div>
-                            </div>
-                            {/* button cập nhật */}
-                            <div className="border-t pt-2 flex justify-end mt-auto">
-                                <button
-                                    onClick={() => setIsType('avatar')}
-                                    className="bg-gray-200 hover:bg-gray-300 rounded-md flex items-center justify-center gap-2 p-2 cursor-pointer mr-2"
-                                >
-                                    <p className="font-semibold">Hủy</p>
-                                </button>
+                        //             {/* Lớp mờ với vùng tròn trong suốt */}
+                        //             <div
+                        //                 className="absolute inset-0 bg-black/50"
+                        //                 style={{
+                        //                     WebkitMaskImage: `radial-gradient(circle 160px at center, transparent 160px, black 161px)`,
+                        //                     maskImage: `radial-gradient(circle 160px at center, transparent 160px, black 161px)`,
+                        //                 }}
+                        //             />
+                        //         </div>
+                        //     </div>
+                        //     {/* button cập nhật */}
+                        //     <div className="border-t pt-2 flex justify-end mt-auto">
+                        //         <button
+                        //             onClick={() => setIsType('avatar')}
+                        //             className="bg-gray-200 hover:bg-gray-300 rounded-md flex items-center justify-center gap-2 p-2 cursor-pointer mr-2"
+                        //         >
+                        //             <p className="font-semibold">Hủy</p>
+                        //         </button>
 
-                                <button
-                                    onClick={onClose}
-                                    className="bg-blue-600 hover:bg-blue-800 rounded-md flex items-center justify-center gap-2 p-2 cursor-pointer mr-3"
-                                >
-                                    <p className="font-semibold text-white">Cập nhật</p>
-                                </button>
-                            </div>
-                        </>
+                        //         <button
+                        //             onClick={onClose}
+                        //             className="bg-blue-600 hover:bg-blue-800 rounded-md flex items-center justify-center gap-2 p-2 cursor-pointer mr-3"
+                        //         >
+                        //             <p className="font-semibold text-white">Cập nhật</p>
+                        //         </button>
+                        //     </div>
+                        // </>
                     )
                 ) : (
                     <ModalAccount setIsType={setIsType} />
-                    // <>
-                    //     <div className="flex-1 overflow-auto relative">
-                    //         {/*Nhấn vào hình bật popup  */}
-                    //         {selectedImage && (
-                    //             <PopupViewImage selectedImage={selectedImage} setSelectedImage={setSelectedImage} />
-                    //         )}
-                    //         <img
-                    //             src={CoverPhoto}
-                    //             alt="Ảnh phụ"
-                    //             className="w-full h-[200px] cursor-pointer"
-                    //             onClick={() => setSelectedImage(CoverPhoto)}
-                    //         />
-
-                    //         <div
-                    //             className="absolute top-44 left-0 w-[80px] h-[80px] rounded-full shadow-lg overflow-hidden ml-6  border-4 border-gray-300 cursor-pointer"
-                    //             onClick={() => setSelectedImage(Avatar)}
-                    //         >
-                    //             <img src={Avatar} alt="Avartar" className="w-full h-full object-cover" />
-                    //         </div>
-
-                    //         <div
-                    //             onClick={() => setIsType('avatar')}
-                    //             className="absolute top-56 left-20 w-8 h-8 bg-gray-200 border border-gray-400 rounded-full flex items-center justify-center cursor-pointer hover:bg-gray-300"
-                    //         >
-                    //             <CiCamera className="w-4 h-4 text-sm" />
-                    //         </div>
-
-                    //         <div
-                    //             className="absolute top-52 left-32 flex items-center justify-center"
-                    //             onClick={() => setIsType('edit')}
-                    //         >
-                    //             <p className="font-semibold mr-3">{userActive?.name}</p>
-                    //             <LuPencilLine className="cursor-pointer" />
-                    //         </div>
-
-                    //         <div className="absolute w-full h-1 bg-gray-200 top-60 mt-6"></div>
-
-                    //         <div className="absolute px-6 top-64 mt-3">
-                    //             <p className="font-semibold">Thông tin cá nhân</p>
-
-                    //             <div className="grid grid-cols-2 gap-x-4 gap-y-2 mt-3 text-sm">
-                    //                 <p className="text-gray-500">Giới tính</p>
-                    //                 <p>{userActive?.gender}</p>
-
-                    //                 <p className="text-gray-500">Ngày sinh</p>
-                    //                 <p>{userActive?.birthDate}</p>
-
-                    //                 <p className="text-gray-500">Điện thoại</p>
-                    //                 <p>(+84 ){userActive?.number}</p>
-                    //             </div>
-
-                    //             <p className="text-gray-500 mt-3 text-xs">
-                    //                 Chỉ bạn bè có thể lưu số của bạn trong danh bạ máy xem được số này
-                    //             </p>
-                    //         </div>
-                    //     </div>
-
-                    //     <div className="border-t pt-2 flex justify-center">
-                    //         <button
-                    //             onClick={() => setIsType('edit')}
-                    //             className="w-11/12 hover:bg-gray-200 rounded-md flex items-center justify-center gap-2 p-2 cursor-pointer"
-                    //         >
-                    //             <LuPencilLine className="cursor-pointer" />
-                    //             <p className="font-semibold">Cập nhật</p>
-                    //         </button>
-                    //     </div>
-                    // </>
                 )}
             </motion.div>
         </div>
